Move `ty::InstantiatedPredicates` to its own little module (cute)
